% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_form_tab.R
\name{flex_form_tab}
\alias{flex_form_tab}
\title{Returns a flexible table with specified parameters}
\usage{
flex_form_tab(
  dates,
  head.df,
  font.size.header = 10,
  font.size.body = 10,
  font.size.total = 10,
  colour.smoll.bord = "gray2",
  colour.big.border = "gray4",
  colour.head.fill = "gray",
  colour.head.text = "black",
  colour.total.fill = "gray",
  nm.total = "Total",
  col.total = 1,
  col.lv1 = NULL,
  lv1.align = "left",
  col.style.prc = NULL,
  col.threshold = NULL,
  colour.fill.threshold = c("olivedrab1", "pink", "grey96"),
  colour.font.threshold = c("darkgreen", "red4", "grey3"),
  threshold = 5,
  inv.threshold = FALSE,
  font.nm = "Century Schoolbook",
  padd = 0,
  bold.head = TRUE,
  bold.body = FALSE,
  bold.total = TRUE
)
}
\arguments{
\item{dates}{\itemize{
\item Data frame
}}

\item{head.df}{\itemize{
\item Table header in the form of a data frame
}}

\item{font.size.header}{\itemize{
\item Header font size
}}

\item{font.size.body}{\itemize{
\item Body font size
}}

\item{font.size.total}{\itemize{
\item Total font size
}}

\item{colour.smoll.bord}{\itemize{
\item Inner border color
}}

\item{colour.big.border}{\itemize{
\item Outer border color
}}

\item{colour.head.fill}{\itemize{
\item Table header color
}}

\item{colour.head.text}{Table header color font}

\item{colour.total.fill}{-Table total color}

\item{nm.total}{\itemize{
\item The word used to mark the total lines
}}

\item{col.total}{\itemize{
\item Column in which to look for total lines
}}

\item{col.lv1}{\itemize{
\item Select a color for the topmost line of a multi-level header
}}

\item{lv1.align}{\itemize{
\item If you change the color of the top line of the header, use alignment
}}

\item{col.style.prc}{\itemize{
\item Columns in which a percent sign is added to the values
}}

\item{col.threshold}{\itemize{
\item Columns for which you need to set a threshold value
}}

\item{colour.fill.threshold}{\itemize{
\item Threshold fill color
}}

\item{colour.font.threshold}{\itemize{
\item Threshold font color
}}

\item{threshold}{\itemize{
\item Threshold value
}}

\item{inv.threshold}{\itemize{
\item FALSE.Invert colors for thresholds
}}

\item{font.nm}{\itemize{
\item Font family
}}

\item{padd}{\itemize{
\item padding
}}

\item{bold.head}{\itemize{
\item TRUE. Table header in bold
}}

\item{bold.body}{\itemize{
\item FALSE. Table body in bold
}}

\item{bold.total}{\itemize{
\item TRUE. Table total in bold
}}
}
\value{
Returns a flexible table with specified parameters
}
\description{
Returns a flexible table with specified parameters
}
