% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_threshold.R
\name{flex_threshold}
\alias{flex_threshold}
\title{Columns for which you need to set a threshold value}
\usage{
flex_threshold(
  ft,
  col.threshold,
  colour.fill.threshold = c("olivedrab1", "pink", "grey96"),
  colour.font.threshold = c("darkgreen", "red4", "grey3"),
  threshold = 5,
  inv.threshold = FALSE
)
}
\arguments{
\item{ft}{\itemize{
\item flex table
}}

\item{col.threshold}{\itemize{
\item Columns for which you need to set a threshold value
}}

\item{colour.fill.threshold}{\itemize{
\item Threshold fill color
}}

\item{colour.font.threshold}{\itemize{
\item Threshold font color
}}

\item{threshold}{\itemize{
\item Threshold value
}}

\item{inv.threshold}{\itemize{
\item FALSE.Invert colors for thresholds
}}
}
\value{
Returns a flexible table with specified parameters
}
\description{
Columns for which you need to set a threshold value
}
