% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tab.R
\name{head_tab}
\alias{head_tab}
\title{Creating a table header}
\usage{
head_tab(dates)
}
\arguments{
\item{dates}{\itemize{
\item Data for creating a table
}}
}
\value{
Table with table header
}
\description{
Creating a table header
}
