% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_flex.R
\name{resize_flex}
\alias{resize_flex}
\title{Resizes flex table to fit powerpoint dimensions}
\usage{
resize_flex(
  ft,
  plan.wid = 12,
  plan.hei = 14,
  col.no = NULL,
  al.w = FALSE,
  cpt = 1
)
}
\arguments{
\item{ft}{\itemize{
\item flex table
}}

\item{plan.wid}{\itemize{
\item Planned width in centimeters
}}

\item{plan.hei}{\itemize{
\item Planned height in centimeters
}}

\item{col.no}{\itemize{
\item Names of columns, except text ones, for which you do not need to resize. Text columns are not resized by default
}}

\item{al.w}{\itemize{
\item FALSE. Automatic alignment of columns, except text and selected ones
}}

\item{cpt}{\itemize{
\item Additional row height scale option. Default is 1.
}}
}
\value{
Returns a flexible table adjusted to the specified dimensions
}
\description{
Resizes flex table to fit powerpoint dimensions
}
