% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tab_pres.R
\name{set_tab_pres}
\alias{set_tab_pres}
\title{Sets options for a flex table}
\usage{
set_tab_pres(
  theme = flextable::theme_vanilla,
  font = "Century Schoolbook",
  mark = " ",
  pd = 1
)
}
\arguments{
\item{theme}{\itemize{
\item Theme (Look flextable)
}}

\item{font}{\itemize{
\item font
}}

\item{mark}{\itemize{
\item mark
}}

\item{pd}{\itemize{
\item padding
}}
}
\value{
\itemize{
\item set
}
}
\description{
Sets options for a flex table
}
