% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_2t_3g.R
\name{slide_2t_3g}
\alias{slide_2t_3g}
\title{Add two tables and a graph}
\usage{
slide_2t_3g(deck, tab1, tab2, gr1, gr2, gr3, sld_loy, name_slide, name_master)
}
\arguments{
\item{deck}{\itemize{
\item presentation
}}

\item{tab1}{\itemize{
\item flexible table
}}

\item{tab2}{\itemize{
\item flexible table
}}

\item{gr1}{\itemize{
\item Graph in format ggplot2
}}

\item{gr2}{\itemize{
\item Graph in format ggplot2
}}

\item{gr3}{\itemize{
\item Graph in format ggplot2
}}

\item{sld_loy}{\itemize{
\item slide layout
}}

\item{name_slide}{\itemize{
\item slide title
}}

\item{name_master}{\itemize{
\item master
}}
}
\value{
Presentation with added slide
}
\description{
Add two tables and a graph
}
