% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_graph.R
\name{slide_graph}
\alias{slide_graph}
\title{Add a table to a slide}
\usage{
slide_graph(deck, gr1, sld_loy, name_slide, name_master)
}
\arguments{
\item{deck}{\itemize{
\item presentation
}}

\item{gr1}{\itemize{
\item flexible table
}}

\item{sld_loy}{\itemize{
\item slide layout
}}

\item{name_slide}{\itemize{
\item slide title
}}

\item{name_master}{\itemize{
\item master
}}
}
\value{
Presentation with added slide
}
\description{
Add a table to a slide
}
