% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_table_graph.R
\name{slide_table_graph}
\alias{slide_table_graph}
\title{Add tables and a graph}
\usage{
slide_table_graph(deck, tab1, gr1, sld_loy, name_slide, name_master)
}
\arguments{
\item{deck}{\itemize{
\item presentation
}}

\item{tab1}{\itemize{
\item flexible table
}}

\item{gr1}{\itemize{
\item Graph in format ggplot2
}}

\item{sld_loy}{\itemize{
\item slide layout
}}

\item{name_slide}{\itemize{
\item slide title
}}

\item{name_master}{\itemize{
\item master
}}
}
\value{
Presentation with added slide
}
\description{
Add tables and a graph
}
