% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_two_table.R
\name{slide_two_table}
\alias{slide_two_table}
\title{Add two tables to a slide}
\usage{
slide_two_table(deck, tab1, tab2, sld_loy, name_slide, name_master)
}
\arguments{
\item{deck}{\itemize{
\item presentation
}}

\item{tab1}{\itemize{
\item flexible table
}}

\item{tab2}{\itemize{
\item flexible table
}}

\item{sld_loy}{\itemize{
\item slide layout
}}

\item{name_slide}{\itemize{
\item slide title
}}

\item{name_master}{\itemize{
\item master
}}
}
\value{
Presentation with added slide
}
\description{
Add two tables to a slide
}
